Refactored slightly to add a core framework
My intention is that Clients should provide instances of Sendable, or
Receivable (when I write it).  The type signatures, and the message
type, are in no way finished yet.

The idea is that Receivable clients will serialize changes (yet to be
written) into reactive-banana events, and Sendable clients will be able
to send messages upstream as per user configuration.

Messages will eventually be renamed into something else, and will be a
typeclass rather than a type.